# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Train Tracking System API
  description: API for managing trains, routes, engines, historical data, and full trains.
  version: 1.0.0
paths:
...
  /routes:
    get:
      summary: Get all routes
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
    post:
      summary: Create a new route
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
        required: true
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
  /routes/{route_id}:
    get:
      summary: Get a specific route by ID
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Route details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
    put:
      summary: Update an existing route
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Route'
        required: true
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
    delete:
      summary: Delete a route
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Route deleted
...
components:
  schemas:
...
    Route:
      type: object
      properties:
        route_id:
          type: string
        route_name:
          type: string
        details:
          type: object
          properties:
            start_station:
              type: string
            end_station:
              type: string
            total_distance_km:
              type: integer
            total_duration_minutes:
              type: integer
      required:
        - route_id
        - route_name
        - details
...